requirements.txt
Django
djangorestframework
drf-yasg
coverage
